blog_backend_cookiecutter/blog_backend_cookiecutter/users/apps.py
import contextlib

from django.apps import AppConfig
from django.utils.translation import gettext_lazy as _


class UsersConfig(AppConfig):
    name = "blog_backend_cookiecutter.users"
    verbose_name = _("Users")

    def ready(self):
        with contextlib.suppress(ImportError):
            import blog_backend_cookiecutter.users.signals  # noqa: F401
